Requirement B1: Implementing a check for attached excursions before deleting vacation

+++ app/src/main/java/edu/wgu/d308pa/fragments/DeleteAlertDialogFragment.java
@@ -9,8 +9,10 @@ import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.DialogFragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+import java.util.List;
 import edu.wgu.d308pa.R;
 import edu.wgu.d308pa.adapters.RecycleViewAdapter;
+import edu.wgu.d308pa.entities.Excursion;
 
 public class DeleteAlertDialogFragment extends DialogFragment {
     @NonNull
@@ -23,14 +25,19 @@ public class DeleteAlertDialogFragment extends DialogFragment {
                 .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        VacationFragment.vacationDao.delete(VacationFragment.vacationDao.findById(vacationId));
 
-                        // refresh the recyclerview
-                        RecyclerView recyclerView = getActivity().findViewById(R.id.vacation_recycler_view);
-                        RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(getContext());
-                        RecycleViewAdapter recycleViewAdapter = new RecycleViewAdapter(getDataForVacationRecyclerView(), getParentFragmentManager());
-                        recyclerView.setAdapter(recycleViewAdapter);
-                        recyclerView.setLayoutManager(layoutManager);
+                        // check for attached excursions
+                        List<Excursion> excursions = VacationFragment.excursionDao.getAllWithVacationId(vacationId);
+                        if (excursions.size() == 0) {
+                            VacationFragment.vacationDao.delete(VacationFragment.vacationDao.findById(vacationId));
+
+                            // refresh the recyclerview
+                            RecyclerView recyclerView = getActivity().findViewById(R.id.vacation_recycler_view);
+                            RecyclerView.LayoutManager layoutManager = new LinearLayoutManager(getContext());
+                            RecycleViewAdapter recycleViewAdapter = new RecycleViewAdapter(getDataForVacationRecyclerView(), getParentFragmentManager());
+                            recyclerView.setAdapter(recycleViewAdapter);
+                            recyclerView.setLayoutManager(layoutManager);
+                        }
                     }
                 })
                 .setNegativeButton("No", new DialogInterface.OnClickListener() {
